+++ src/server/scripts/Northrend/FrozenHalls/HallsOfReflection/boss_marwyn.cpp
@@ -82,7 +82,7 @@ public:
 
        uint8 SummonCount;
 
-       uint64 pSummon;
+       uint32 pSummon;
 
        void Reset()
        {
